Extract TimeSlice class.
